requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.28.1
langgraph>=0.0.62
langchain-core>=0.2.0
langchain>=0.3.0
langchain-openai>=0.0.2
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
pydantic>=2.5.0
requests>=2.31.0